requirements.txt
acres==0.5.0
altair==5.5.0
altgraph==0.17.4
attrs==24.2.0
blinker==1.8.2
bottle==0.13.4
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cx_Freeze==8.4.1
cycler==0.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
docx==0.2.4
et-xmlfile==1.1.0
etelemetry==0.3.1
filelock==3.19.1
fitz==0.0.1.dev2
fonttools==4.55.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.1.1
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lab==8.5
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.3.2
patchelf==0.17.2.4
pathlib==1.0.1
pdfkit==1.0.0
pillow==10.4.0
plotly==6.3.0
protobuf==5.28.3
prov==2.1.1
proxy_tools==0.1.0
puremagic==1.30
pyarrow==18.0.0
pydeck==0.9.1
pydot==4.0.1
Pygments==2.18.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyserial==3.5
python-can==4.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2024.1
pywebview==5.4
pyxnat==1.6.3
rdflib==7.2.1
referencing==0.35.1
regex==2025.10.22
report==0.0.1
reportlab==4.4.4
requests==2.32.3
retrying==1.3.4
rpds-py==0.20.0
scipy==1.16.2
setuptools==80.9.0
simplejson==3.20.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
streamlit==1.50.0
streamlit-aggrid==1.1.9
streamlit-calendar==1.4.0
streamlit-desktop-app==0.3.4
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traits==7.0.2
txt2tags==3.9
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.6
wrapt==1.17.3
zipp==3.20.2
